pyproject.toml
[project]
name = "prompt-wrangler"
version = "0.1.0"
description = "NER keyword extraction prompt playground for medical texts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "instructor>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
prompt-wrangler = "app.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
